bu_processor/requirements.txt
# ============================================================================
# BU-PROCESSOR REQUIREMENTS
# ============================================================================
# Core dependencies for the BU-Processor pipeline
# Compatible with Python 3.8+

# ============================================================================
# MACHINE LEARNING & TRANSFORMERS
# ============================================================================
transformers>=4.40.0
sentence-transformers>=2.7.0
torch>=2.3.0
scikit-learn>=1.4.0
datasets>=2.19.0

# ============================================================================
# DATA SCIENCE & NUMERIC LIBRARIES  
# ============================================================================
numpy>=1.26.0
pandas>=2.2.0
mmh3>=4.0.1

# ============================================================================
# PDF PROCESSING
# ============================================================================
PyMuPDF>=1.24.5
PyPDF2>=3.0.1
PDFplumber>=0.11.0
reportlab>=4.2.0

# ============================================================================
# WEB & API FRAMEWORKS
# ============================================================================
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.6

# ============================================================================
# CONFIGURATION & VALIDATION
# ============================================================================
pydantic>=2.7.0
pydantic-settings>=2.2
python-dotenv>=1.0.1

# ============================================================================
# DATABASE & STORAGE
# ============================================================================
sqlalchemy>=2.0

# ============================================================================
# VECTOR DATABASE
# ============================================================================
pinecone>=5.0.0
faiss-cpu>=1.7.4              # For local vector index (alternative to pinecone)
rank-bm25>=0.2.2               # For BM25 lexical search

# ============================================================================
# CHATBOT INTEGRATION (OPTIONAL)
# ============================================================================
openai>=1.35.0

# ============================================================================
# CLI & USER EXPERIENCE
# ============================================================================
typer>=0.12.0
rich>=13.7.1
tqdm>=4.66.4

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
structlog>=24.1.0
prometheus_client>=0.20.0

# ============================================================================
# VISUALIZATION (OPTIONAL)
# ============================================================================
matplotlib>=3.9.0
plotly>=5.22.0
seaborn>=0.13.2

# ============================================================================
# UTILITIES & HELPERS
# ============================================================================
backoff>=2.2.1
aiohttp>=3.9.0

# ============================================================================
# DEVELOPMENT DEPENDENCIES (install separately with requirements-dev.txt)
# ============================================================================
# pytest>=8.0.0
# pytest-cov>=4.0.0
# black>=24.0.0
# isort>=5.13.0
# mypy>=1.8.0
# flake8>=7.0.0
# pre-commit>=3.6.0
